requirement.txt
selenium
unittest
allure-pytest
pytest